Check if object is any content type

Internal helper to check if an object is any type of content, including both content collections and individual content blocks.

Usage

is_any_content(x)

Arguments

x

Object to check

Value

Logical indicating if x is any content type